+++ src/pages/login/Interface/Interface.ts
@@ -1,0 +1,9 @@
+interface LoginInterface {
+  message: string;
+  email: string;
+  user_id: string;
+  name: string;
+  isFaceVerified: boolean;
+}
+
+export type { LoginInterface };
