+++ src/pages/signup/Interface/interface.ts
@@ -1,0 +1,4 @@
+interface SignUpInterface {
+  message: string;
+}
+export type { SignUpInterface };
